add navbar functionality
